requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pdfplumber==0.11.0
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
